% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get}
\alias{data_get}
\title{Return a dataset}
\usage{
data_get(dataset)
}
\arguments{
\item{dataset}{as character}
}
\description{
Return a dataset
}
\examples{
data_get("all-permits")
}
\keyword{datasets}

