% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_departments.R
\name{list_departments}
\alias{list_departments}
\title{List all departments}
\usage{
list_departments(search = NULL)
}
\arguments{
\item{search}{Optional string to search by as character}
}
\description{
List all departments
}
\examples{
list_departments()
list_departmetns("Louisville")
}
\keyword{departments}

