% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_revisions.R
\name{list_revisions}
\alias{list_revisions}
\title{List all revisions}
\usage{
list_revisions()
}
\arguments{
\item{no}{parameters}
}
\description{
List all revisions
}
\examples{
list_revisions()
}
\keyword{groups}

